add validator

+++ src/main/java/com/ua/sutty/spring/controller/DeleteController.java
@@ -1,7 +1,9 @@
 package com.ua.sutty.spring.controller;
 
+import com.ua.sutty.spring.security.details.UserDetailsImpl;
 import com.ua.sutty.spring.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -13,7 +15,10 @@ public class DeleteController {
     private UserService userService;
 
     @GetMapping("/delete/{id}")
-    public String delete(@PathVariable Long id){
+    public String delete(@PathVariable Long id, Authentication authentication){
+        if (userService.findUserById(id).equals(((UserDetailsImpl) authentication.getPrincipal()).getUser())){
+            return "redirect:/home?deleteYourself";
+        }
         userService.deleteUserById(id);
         return "redirect:/home?successfullyDeleted";
     }
